// ui/src/api/types.gen.ts
export type DtoViewFilters = {
    assignee_id?: string;
    cycle_id?: string;
    label_ids?: Array<string>;
    priority?: string;
    search?: string;
    status_id?: string;
};

export type DtoViewResponse = {
    created_at?: string;
    created_by?: string;
    filters?: DtoViewFilters;
    id?: string;
    layout?: string;
    name?: string;
    project_id?: string;
    project_name?: string;
    project_slug?: string;
    updated_at?: string;
};